requirements.txt
requests>=2.26.0
beautifulsoup4>=4.9.3
openai>=1.0.0,<2.0.0
python-dotenv>=0.19.0
pyyaml==6.0.1
pyairtable==2.0.0
schedule==1.2.0
loguru==0.7.2
selenium>=4.15.2
webdriver-manager>=4.0.1
black==24.2.0
flake8==6.0.0
pygithub==2.1.1